requirements.txt
pyodbc
tabulate
openai
langchain
langchain-community
langchain-core
langchain-experimental
langchain-openai
langchain-groq
SQLAlchemy
pandas
python-dotenv
streamlit